First attempt at JSON output. Can output posts, but haven't figured out how to customize endpoints to include tags, categories and full url of featured image.

+++ app/Http/Controllers/Dashboard/AdminPostsController.php
@@ -130,7 +130,9 @@ class AdminPostsController extends Controller
     {
         $post = Post::findOrFail($id);
 
-        unlink(public_path() . '/images/' . $post->photo->file);
+        if($post->photo->file) {
+             unlink(public_path() . '/images/' . $post->photo->file);
+        }
 
         $post->delete();
